[Prover] Rename config.rs -> prover_config.rs
