+++ prover-service/src/lib.rs
@@ -3,7 +3,6 @@
 extern crate core;
 
 pub mod api;
-pub mod config;
 pub mod deployment_information;
 pub mod error;
 pub mod groth16_vk;
@@ -13,6 +12,7 @@ pub mod jwk_fetching;
 pub mod load_vk;
 pub mod logging;
 pub mod metrics;
+pub mod prover_config;
 pub mod prover_key;
 pub mod proving;
 pub mod state;
